% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ch_dl}
\alias{ch_dl}
\title{Download CHELSA data}
\usage{
ch_dl(md, dest = NULL, skip_existing = TRUE, method = "curl",
  crop = NULL)
}
\arguments{
\item{md}{Variables to download (a data frame created by ch_queries).}

\item{dest}{Path to file folder where downloaded files should be stored
(character).}

\item{skip_existing}{Should files that are already present in the destination
be igored (logical).}

\item{method}{Download method, passed to downloa.file (character).}

\item{crop}{Spatial bounding box to crop each downloaded raster to (an extent
object, or any spatial object with an extent).}
}
\description{
Many standard palette generators use only a slice of color space, which can
cause a lack of differentiability in palettes used to visualize categorical
factors with many levels. This function attempts to overcome this by
generating colors using nearest-neighbor distance maximization in 3D RGB
space.
}
