% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ch_queries}
\alias{ch_queries}
\title{Build a list of files to download}
\usage{
ch_queries(variables, months, models, scenarios, timeframes)
}
\arguments{
\item{variables}{character vector.}

\item{months}{character vector.}

\item{models}{character vector.}

\item{scenarios}{character vector.}

\item{timeframes}{character vector.}
}
\value{
a data frame of metadata for all factorial combinations of the
  requested variables
}
\description{
This function constructs a set of download URLs for the requested
combinations of metadata variables
}
